Cauldron of desire logic
He works!! Next up I just need to transform him into a proper boss with cool actions and a better, larger sprite and hitbox.

+++ Assets/Source/Enemies/FiniteStateMachine/FloorBosses/Cauldron of Desire/COD_AllCardsDrawn.cs
@@ -17,7 +17,7 @@ namespace Cardificer.FiniteStateMachine
         {
             var cardsDrawnExists = state.trackedVariables.TryGetValue("CardsDrawn", out var cardsDrawn);
             var cardsToDrawExists = state.trackedVariables.TryGetValue("CardsToDraw", out var cardsToDraw);
-            return cardsDrawnExists && cardsToDrawExists && cardsDrawn == cardsToDraw;
+            return cardsDrawnExists && cardsToDrawExists && (int)cardsDrawn == (int)cardsToDraw;
         }
     }
 }
